requirements.txt
Flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
google-generativeai==0.8.4
requests==2.32.3
Werkzeug==3.1.3
urllib3==2.3.0
gevent==24.2.1
